+++ extensions/vscode/src/extension.ts
@@ -17,6 +17,7 @@ import { HomeViewProvider } from "src/views/homeView";
 import { WatcherManager } from "src/watchers";
 import { Commands } from "src/constants";
 import { DocumentTracker } from "./entrypointTracker";
+import { getXDGConfigProperty } from "src/utils/config";
 
 const STATE_CONTEXT = "posit.publish.state";
 
@@ -171,9 +172,17 @@ export const extensionSettings = {
   },
   defaultConnectServer(): string {
     const configuration = workspace.getConfiguration("positPublisher");
-    const value: string | undefined = configuration.get<string>(
+    let value: string | undefined = configuration.get<string>(
       "defaultConnectServer",
     );
+
+    if (value === undefined || value === "") {
+      const configURL: string | null = getXDGConfigProperty(
+        "rstudio/rsession.conf",
+        "default-rsconnect-server",
+      );
+      if (configURL !== null) value = configURL;
+    }
     return value !== undefined ? value : "";
   },
 };